Split dialog into core components
