+++ src/components/admin/Settings/StorageMappings/Dialog/useRepublishPrefix.ts
@@ -55,6 +55,7 @@ function useRepublishPrefix() {
             const evaluatedPubId = republicationResponse.data;
             setPubId(evaluatedPubId);
 
+            // TODO (storageMappings): Double check that this call is actually retried.
             const publicationResponse = await supabaseRetry(
                 () => getPublicationByIdQuery(evaluatedPubId),
                 'getPublicationById'
